backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
fastapi-socketio
redis
asyncssh
httpx
